+++ Tests/KeyboardKitSwiftUITests/Extensions/Color+ButtonTests.swift
@@ -1,5 +1,5 @@
 //
-//  Color+SystemTests.swift
+//  Color+ButtonTests.swift
 //  KeyboardKitTests
 //
 //  Created by Daniel Saidi on 2020-06-24.
@@ -12,7 +12,7 @@ import KeyboardKit
 import SwiftUI
 @testable import KeyboardKitSwiftUI
 
-class Color_SystemTests: QuickSpec {
+class Color_ButtonTests: QuickSpec {
 
     override func spec() {
         
@@ -25,7 +25,18 @@ class Color_SystemTests: QuickSpec {
         
         func result(for color: Color) -> Color { color }
         
-        describe("dark system keyboard button background color") {
+        describe("standard button shadow color") {
+            
+            it("is 30% black") {
+                let expected = Color.black.opacity(0.3)
+                expect(result(for: .standardButtonShadowColor(for: context(.dark, .dark)))).to(equal(expected))
+                expect(result(for: .standardButtonShadowColor(for: context(.dark, .light)))).to(equal(expected))
+                expect(result(for: .standardButtonShadowColor(for: context(.light, .dark)))).to(equal(expected))
+                expect(result(for: .standardButtonShadowColor(for: context(.light, .light)))).to(equal(expected))
+            }
+        }
+        
+        describe("standard dark button background color") {
             
             it("varies for different schemes and appearances") {
 //                expect(result(for: .systemKeyboardButtonBackgroundColorDark(for: context(.dark, .dark)))).to(equal(.systemKeyboardButtonBackgroundColorDarkForDarkColorScheme))
@@ -35,7 +46,7 @@ class Color_SystemTests: QuickSpec {
             }
         }
         
-        describe("light system keyboard button background color") {
+        describe("standard light button background color") {
             
             it("varies for different schemes and appearances") {
 //                expect(result(for: .systemKeyboardButtonBackgroundColorLight(for: context(.dark, .dark)))).to(equal(.systemKeyboardButtonBackgroundColorLightForDarkColorScheme))
@@ -45,15 +56,15 @@ class Color_SystemTests: QuickSpec {
             }
         }
         
-        describe("system keyboard button background color") {
+        describe("standard button background color") {
             
             it("varies for different schemes and appearances") {
-                expect(result(for: .systemKeyboardButtonBackgroundColorDarkForDarkColorScheme)).to(equal(Color(rgb: 71.0/255.0)))
-                expect(result(for: .systemKeyboardButtonBackgroundColorDarkForLightColorSchemeAndDarkKeyboardAppearance)).to(equal(Color(rgb: 117.0/255.0)))
-                expect(result(for: .systemKeyboardButtonBackgroundColorDarkForLightColorSchemeAndLightKeyboardAppearance)).to(equal(Color(red: 179.0/255.0, green: 184.0/255.0, blue: 193.0/255.0)))
-                expect(result(for: .systemKeyboardButtonBackgroundColorLightForDarkColorScheme)).to(equal(Color(rgb: 107.0/255.0)))
-                expect(result(for: .systemKeyboardButtonBackgroundColorLightForLightColorSchemeAndDarkKeyboardAppearance)).to(equal(Color(rgb: 150.0/255.0)))
-                expect(result(for: .systemKeyboardButtonBackgroundColorLightForLightColorSchemeAndLightKeyboardAppearance)).to(equal(.white))
+                expect(result(for: .standardDarkButtonBackgroundColorForDarkColorScheme)).to(equal(Color(rgb: 71.0/255.0)))
+                expect(result(for: .standardDarkButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance)).to(equal(Color(rgb: 117.0/255.0)))
+                expect(result(for: .standardDarkButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance)).to(equal(Color(red: 179.0/255.0, green: 184.0/255.0, blue: 193.0/255.0)))
+                expect(result(for: .standardLightButtonBackgroundColorForDarkColorScheme)).to(equal(Color(rgb: 107.0/255.0)))
+                expect(result(for: .standardLightButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance)).to(equal(Color(rgb: 150.0/255.0)))
+                expect(result(for: .standardLightButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance)).to(equal(.white))
             }
         }
         
@@ -64,17 +75,6 @@ class Color_SystemTests: QuickSpec {
 //                expect(result(for: .systemKeyboardButtonForegroundColor(for: context(.dark, .light)))).to(equal(.primary))
 //                expect(result(for: .systemKeyboardButtonForegroundColor(for: context(.light, .dark)))).to(equal(.white))
 //                expect(result(for: .systemKeyboardButtonForegroundColor(for: context(.light, .light)))).to(equal(.primary))
-            }
-        }
-        
-        describe("system keyboard button shadow color") {
-            
-            it("is 30% black") {
-                let expected = Color.black.opacity(0.3)
-                expect(result(for: .systemKeyboardButtonShadowColor(for: context(.dark, .dark)))).to(equal(expected))
-                expect(result(for: .systemKeyboardButtonShadowColor(for: context(.dark, .light)))).to(equal(expected))
-                expect(result(for: .systemKeyboardButtonShadowColor(for: context(.light, .dark)))).to(equal(expected))
-                expect(result(for: .systemKeyboardButtonShadowColor(for: context(.light, .light)))).to(equal(expected))
             }
         }
     }